src/src_buildroot/buildroot/ATC_PACKAGES/atcchessenginetheolepage/src/tests/requirements.txt
python-chess
termcolor